requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
passlib[argon2]
structlog
redis
httpx
pytest
pytest-asyncio
pytest-cov
aiosqlite
python-multipart
email-validator
argon2-cffi
cryptography>=41.0.0
bleach

